Move some tests from `integration` to `unit`
