+++ tests/integration/test___main__.py
@@ -6,19 +6,7 @@ from pathlib import Path
 
 import pytest
 
-from ehrql.__main__ import (
-    BACKEND_ALIASES,
-    QUERY_ENGINE_ALIASES,
-    backend_from_id,
-    main,
-    query_engine_from_id,
-)
-from ehrql.backends.base import SQLBackend
-from ehrql.query_engines.base import BaseQueryEngine
-from ehrql.query_engines.base_sql import BaseSQLQueryEngine
-from ehrql.query_engines.debug import DebugQueryEngine
-from ehrql.query_engines.in_memory import InMemoryQueryEngine
-from ehrql.utils.module_utils import get_sibling_subclasses
+from ehrql.__main__ import main
 from tests.lib.inspect_utils import function_body_as_string
 
 
@@ -165,36 +153,3 @@ def test_graph_query(tmpdir):  # pragma: no cover
         ]
     )
     assert output_file.exists()
-
-
-def test_all_query_engine_aliases_are_importable():
-    for alias in QUERY_ENGINE_ALIASES.keys():
-        assert query_engine_from_id(alias)
-
-
-def test_all_backend_aliases_are_importable():
-    for alias in BACKEND_ALIASES.keys():
-        assert backend_from_id(alias)
-
-
-def test_all_query_engines_have_an_alias():
-    for cls in get_sibling_subclasses(BaseQueryEngine):
-        if cls in [
-            BaseSQLQueryEngine,
-            InMemoryQueryEngine,
-            DebugQueryEngine,
-        ]:
-            continue
-        name = f"{cls.__module__}.{cls.__name__}"
-        assert name in QUERY_ENGINE_ALIASES.values(), f"No alias defined for '{name}'"
-
-
-def test_all_backends_have_an_alias():
-    for cls in get_sibling_subclasses(SQLBackend):
-        name = f"{cls.__module__}.{cls.__name__}"
-        assert name in BACKEND_ALIASES.values(), f"No alias defined for '{name}'"
-
-
-def test_all_backend_aliases_match_display_names():
-    for alias in BACKEND_ALIASES.keys():
-        assert backend_from_id(alias).display_name.lower() == alias
