+++ tests/unit/test___main__.py
@@ -1,6 +1,8 @@
 import pytest
 
 from ehrql.__main__ import (
+    BACKEND_ALIASES,
+    QUERY_ENGINE_ALIASES,
     ArgumentTypeError,
     DefinitionError,
     FileValidationError,
@@ -9,6 +11,12 @@ from ehrql.__main__ import (
     main,
     query_engine_from_id,
 )
+from ehrql.backends.base import SQLBackend
+from ehrql.query_engines.base import BaseQueryEngine
+from ehrql.query_engines.base_sql import BaseSQLQueryEngine
+from ehrql.query_engines.debug import DebugQueryEngine
+from ehrql.query_engines.in_memory import InMemoryQueryEngine
+from ehrql.utils.module_utils import get_sibling_subclasses
 from tests.lib.inspect_utils import function_body_as_string
 
 
@@ -283,6 +291,39 @@ def test_backend_from_id_wrong_type():
 @pytest.mark.parametrize("alias", ["expectations", "test"])
 def test_backend_from_id_special_case_aliases(alias):
     assert backend_from_id(alias) is None
+
+
+def test_all_query_engine_aliases_are_importable():
+    for alias in QUERY_ENGINE_ALIASES.keys():
+        assert query_engine_from_id(alias)
+
+
+def test_all_backend_aliases_are_importable():
+    for alias in BACKEND_ALIASES.keys():
+        assert backend_from_id(alias)
+
+
+def test_all_query_engines_have_an_alias():
+    for cls in get_sibling_subclasses(BaseQueryEngine):
+        if cls in [
+            BaseSQLQueryEngine,
+            InMemoryQueryEngine,
+            DebugQueryEngine,
+        ]:
+            continue
+        name = f"{cls.__module__}.{cls.__name__}"
+        assert name in QUERY_ENGINE_ALIASES.values(), f"No alias defined for '{name}'"
+
+
+def test_all_backends_have_an_alias():
+    for cls in get_sibling_subclasses(SQLBackend):
+        name = f"{cls.__module__}.{cls.__name__}"
+        assert name in BACKEND_ALIASES.values(), f"No alias defined for '{name}'"
+
+
+def test_all_backend_aliases_match_display_names():
+    for alias in BACKEND_ALIASES.keys():
+        assert backend_from_id(alias).display_name.lower() == alias
 
 
 def test_debug(capsys, tmp_path):
